fix: Renamed Duration to TickDuration
