+++ src/bsgo/time/CMakeLists.txt
@@ -1,7 +1,7 @@
 
 target_sources (bsgo_lib PRIVATE
-	${CMAKE_CURRENT_SOURCE_DIR}/Duration.cc
 	${CMAKE_CURRENT_SOURCE_DIR}/Tick.cc
+	${CMAKE_CURRENT_SOURCE_DIR}/TickDuration.cc
 	)
 
 target_include_directories (bsgo_lib PUBLIC
