+++ tests/unit/bsgo/time/CMakeLists.txt
@@ -4,6 +4,6 @@ target_include_directories(unitTests PUBLIC
 	)
 
 target_sources(unitTests PUBLIC
-	${CMAKE_CURRENT_SOURCE_DIR}/DurationTest.cc
+	${CMAKE_CURRENT_SOURCE_DIR}/TickDurationTest.cc
 	${CMAKE_CURRENT_SOURCE_DIR}/TickTest.cc
 	)
